requirements.txt
crewai
fastapi
uvicorn[standard]
python-dotenv
kafka-python
redis
langchain